IndexBuffer 를 이용한 DrawIndexed 출력
+ GameProcess.cpp Line 112 에러로 인한 비활성화

+++ GA2023_3Q_DX/DX_Tutorials_SH/Basic/GameProcess.h
@@ -41,7 +41,7 @@ private:
 
 	// IA
 	vector<Vertex> m_vertices;
-	vector<size_t> m_indices;
+	vector<WORD> m_indices;
 
 	ID3D11Buffer* m_vertexBuffer = nullptr;
 	ID3D11Buffer* m_indexBuffer = nullptr;
